14 and 15 updated

+++ 14-binary_tree_balance.c
@@ -35,7 +35,7 @@ size_t binary_tree_height(const binary_tree_t *tree)
  * Return: the number of leaves in the tree.
  */
 
-size_t binary_tree_balance(const binary_tree_t *tree)
+int binary_tree_balance(const binary_tree_t *tree)
 {
 	int left_height = 0, right_height = 0;
